replace variables to util file
