+++ app/src/main/java/com/example/mbanking/util/Variables.kt
@@ -1,0 +1,9 @@
+package com.example.mbanking.util
+
+import com.example.mbanking.data.AccountData
+
+var newStartDate = "01.01.1970"
+var newEndDate = "30.01.3000"
+var listOfAccounts = mutableListOf(AccountData.First, AccountData.Second, AccountData.Third)
+var transactionIter = 0
+var accountValue: Int = 0
